pyproject.toml
[build-system]
requires = [
    "setuptools",
    "wheel"
]
build-backend = "setuptools.build_meta"

[project]
name = "apache-airflow-providers-pysparkonk8s"
dynamic = [
    "version"
]
description = "Provider package apache-airflow-providers-pysparkonk8s for Apache Airflow"
authors = [
    { name="Sebastian Daberdaku", email="sebastiandaberdaku@gmail.com" },
]
maintainers = [
    { name="Sebastian Daberdaku", email="sebastiandaberdaku@gmail.com" },
]
license = { text = "Apache License 2.0" }
keywords = [ "airflow-provider", "airflow", "spark", "pyspark", "kubernetes", "integration" ]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Environment :: Web Environment",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "Framework :: Apache Airflow",
    "Framework :: Apache Airflow :: Provider",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Monitoring",
]
requires-python = "~=3.10"
dependencies = [
	"apache-airflow>=2.6.0",
	"apache-airflow-providers-cncf-kubernetes>=7.4.0",
	"PyJWT==2.8.0",
    "kubernetes~=23.6.0",
    "pyyaml~=6.0",
	"pyspark>=3.5.0",
]

[project.urls]
"Source Code" = "https://github.com/sebastiandaberdaku/apache-airflow-providers-pysparkonk8s"

[project.readme]
file = "README.md"
content-type = "text/markdown"

[project.entry-points."apache_airflow_provider"]
provider_info = "airflow.providers.pysparkonk8s.get_provider_info:get_provider_info"

[tool.setuptools]
include-package-data = false

[tool.setuptools.dynamic]
version = { attr = "airflow.providers.pysparkonk8s.get_provider_info.__version__" }

[tool.setuptools.packages.find]
include = [
    "airflow.providers.*"
]
exclude = [
    "tests",
    "tests*",
    "examples"
]